pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "screencast-narrator"
version = "0.1.0"
description = "Turn raw screen recordings + narration text into polished narrated screencasts"
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Martin Marinschek" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "Pillow>=10.0",
    "qrcode>=7.4",
    "pyzbar>=0.1.9",
]

[project.optional-dependencies]
tts = [
    "kokoro>=0.3",
    "soundfile>=0.12",
    "numpy>=1.24",
]
dev = [
    "pytest>=7.0",
]

[project.scripts]
screencast-narrator = "screencast_narrator.merge:main"

[project.urls]
Homepage = "https://github.com/mmarinschek/screencast-narrator"
Repository = "https://github.com/mmarinschek/screencast-narrator"
Issues = "https://github.com/mmarinschek/screencast-narrator/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
